[refactor] 프로필 이미지 패키지 분리 #16
