+++ src/test/java/com/triptune/domain/member/controller/MemberControllerTest.java
@@ -7,9 +7,9 @@ import com.triptune.domain.member.dto.request.ChangeNicknameRequest;
 import com.triptune.domain.member.dto.request.ChangePasswordRequest;
 import com.triptune.domain.member.dto.request.JoinRequest;
 import com.triptune.domain.member.entity.Member;
-import com.triptune.domain.member.entity.ProfileImage;
+import com.triptune.domain.profile.entity.ProfileImage;
 import com.triptune.domain.member.repository.MemberRepository;
-import com.triptune.domain.member.repository.ProfileImageRepository;
+import com.triptune.domain.profile.repository.ProfileImageRepository;
 import com.triptune.global.enumclass.ErrorCode;
 import com.triptune.global.enumclass.SuccessCode;
 import com.triptune.global.util.JwtUtil;
