+++ src/test/java/com/triptune/domain/schedule/controller/ScheduleControllerTest.java
@@ -3,9 +3,9 @@ package com.triptune.domain.schedule.controller;
 import com.triptune.domain.common.entity.*;
 import com.triptune.domain.common.repository.*;
 import com.triptune.domain.member.entity.Member;
-import com.triptune.domain.member.entity.ProfileImage;
+import com.triptune.domain.profile.entity.ProfileImage;
 import com.triptune.domain.member.repository.MemberRepository;
-import com.triptune.domain.member.repository.ProfileImageRepository;
+import com.triptune.domain.profile.repository.ProfileImageRepository;
 import com.triptune.domain.schedule.ScheduleTest;
 import com.triptune.domain.schedule.dto.request.ScheduleCreateRequest;
 import com.triptune.domain.schedule.dto.request.RouteRequest;
