+++ src/test/java/com/triptune/domain/schedule/service/ScheduleServiceTest.java
@@ -2,7 +2,7 @@ package com.triptune.domain.schedule.service;
 
 import com.triptune.domain.common.entity.*;
 import com.triptune.domain.member.entity.Member;
-import com.triptune.domain.member.entity.ProfileImage;
+import com.triptune.domain.profile.entity.ProfileImage;
 import com.triptune.domain.member.repository.MemberRepository;
 import com.triptune.domain.schedule.ScheduleTest;
 import com.triptune.domain.schedule.dto.*;
